Lixing

+++ boundary/agent_manageCreatedListings.php
@@ -1,27 +1,27 @@
 <?php 
 require_once "partials/header.php"; 
-require_once "../controller/agentViewListingController.php";
+require_once "../controller/agentViewCreatedListingController.php";
 require_once "../controller/agentSearchListingController.php";
 require_once "../controller/agentDeleteListingController.php";
 
 $loggedInUsername = $_SESSION["username"];
-$agentViewListingController = new agentViewListingController();
+$agentViewAllListingController = new AgentViewCreatedListingController();
 $allListing;
-$errors;
 
 // display new listings
 function displayCreatedListings()
 {
     global $loggedInUsername;
-    global $agentViewListingController;
+    global $agentViewAllListingController;
     global $allListing;
 
-    $allListing = $agentViewListingController->agentGetCreatedListings($loggedInUsername);  
+    $allListing = $agentViewAllListingController->agentGetCreatedListings($loggedInUsername);  
 }
 
 function searchCreatedListings()
 {
     global $allListing;
+    global $loggedInUsername;
 
     $searchInfo = array();
 
@@ -30,11 +30,13 @@ function searchCreatedListings()
         $searchInfo[$key] = $value;
     }
 
+    $searchInfo['listed_by'] = $loggedInUsername;
+
     $agentSearchListingController = new AgentSearchListingController();
     $allListing = $agentSearchListingController->agentSearchListings($searchInfo);
 }
 
-function deleteCreatedListings()
+function deleteAListing()
 {
     $agentDeleteListingController = new AgentDeleteListingController();
     $deleted = $agentDeleteListingController->agentDeleteListings($_GET["delete_id"]);
@@ -57,7 +59,7 @@ function deleteSuccess()
         </div>';
 
     // Redirect using JavaScript after the alert is closed
-    echo '<script>setTimeout(function() { window.location.href = "agent_manageListings.php"; }, 1000);</script>';
+    echo '<script>setTimeout(function() { window.location.href = "agent_manageCreatedListings.php"; }, 1000);</script>';
 }
 
 function deleteFail()
@@ -66,8 +68,6 @@ function deleteFail()
             Delete fail..try again!
         </div>';
 }
-
-
 
 // search listings created by agent
 if(isset($_GET['searchForm']))
@@ -81,20 +81,23 @@ else
 
 if(isset($_GET['delete_id']))
 {
-    deleteCreatedListings();
+    deleteAListing();
 }
 
 ?>
+
+<br>
+<h2> &nbsp;My Listed Properties</h2>
 
 <!-- search bar -->
 <div class="container-fluid mt-3">
     <form class="form-inline" method="GET" action=""
             style="background-color: grey; padding: 10px; border-radius: 5px; width: 90vw;">
         <input class="form-control mr-sm-2" type="search" placeholder="Title/type/location/status/seller" aria-label="Search" name="search" style="width: 25%;">
-        <input class="form-control mr-sm-2" type="number" placeholder="Min Price" name="min_price" style="width: 15%;">
-        <input class="form-control mr-sm-2" type="number" placeholder="Max Price" name="max_price" style="width: 15%;">
-        <input class="form-control mr-sm-2" type="number" placeholder="Min Area" name="min_area" style="width: 15%;">
-        <input class="form-control mr-sm-2" type="number" placeholder="bedroom+hall+kitchen num" name="bhk" style="width: 20;">
+        <input class="form-control mr-sm-2" type="number" placeholder="Min Price" name="min_price" style="width: 15%;" min="0">
+        <input class="form-control mr-sm-2" type="number" placeholder="Max Price" name="max_price" style="width: 15%;" min="0">
+        <input class="form-control mr-sm-2" type="number" placeholder="Min Area" name="min_area" style="width: 15%;" min="0">
+        <input class="form-control mr-sm-2" type="number" placeholder="bedroom+hall+kitchen num" name="bhk" style="width: 20;" min="0">
         <button class="btn btn-success my-2 my-sm-0" type="submit" name="searchForm" value="search" style="width: 10%;">
             <i class="fas fa-search"></i> Search
         </button>
@@ -148,7 +151,7 @@ if (empty($allListing)) {
                         <i class="fas fa-eye"></i>View</a>
                         <a href="agent_updateListings.php?listing_id=<?php echo $listing['listing_id']; ?>" class="btn btn-success btn-sm">
                         <i class="fas fa-pencil"></i>Update</a>
-                        <a href="?delete_id=<?php echo $listing['listing_id']; ?>" class="btn btn-danger btn-sm">
+                        <a href="?delete_id=<?php echo $listing['listing_id']; ?>" class="btn btn-danger btn-sm" onclick="return confirm('confirm delete')">
                         <i class="fas fa-trash"></i>Delete</a>
                     </td>
                 </tr>
